Written by: Landry Breuil <landry@openbsd.org>
Import sdlmame 0.130 :
MAME stands for Multi Arcade Machine Emulator. As the name suggests, it
is a program that emulates the hardware (and low-level firmware) of a
massive variety of arcade machines.
SDLMAME is a port of MAME using the SDL library.

After many submissions and updates by Pierre Riteau (at gmail), thanks!