Written by: Landry Breuil <landry@openbsd.org>
Import roadfighter, a remake of the original game from Konami in 1985.

In Road Fighter you drive a car in a death race between you and a
group of mad drivers. Your aim? To reach the end of the checkpoint
before you run out of fuel. Sounds easy? Far from it, there's a lot of
obstacles between your car and the checkpoints!