Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
Rework the way we handle build_depends with the gnome MODULE when
MODGNOME_HELP_FILES is set to yes.

* add a build_depends on gnome-doc-utils if MODGNOME_HELP_FILES is set
to yes
* add rarian as a run_depends for gnome-doc-utils (will help with older
ports that will never be updated from using deprecated scrollkeeper)
* remove build+run_depends on gnome-doc-utils and rarian for ports that
have MODGNOME_HELP_FILES as they are now implied.
* bump where needed

This should help in having all required build dependencies for
gnome-related ports and it greatly simplifies things.

"thanks for cleaning this mess" and ok jasper@