Written by: Landry Breuil <landry@openbsd.org>
Import xfce4-notifyd, a simple alternative notification daemon for Xfce,
following the fd.o desktop notification spec. In the long term, it will
supersede notification-daemon-xfce, which was only a de-gnomified version
of notification-daemon, and is unmaintained upstream as of now.
Ah, and it ships an awesome ZOMG-PONIES! theme, which makes this package
totally unique in the ports-tree.