Written by: Simon Bertrang <simon@openbsd.org>
import p5-YAML-Tiny

The YAML specification is huge.  Really, really huge.  It contains all
the functionality of XML, except with flexibility and choice, which
makes it easier to read, but with a formal specification that is more
complex than XML.
The original pure-Perl implementation YAML costs just over 4 megabytes
of memory to load.  Just like with Windows .ini files (3 meg to load)
and CSS (3.5 meg to load) the situation is just asking for a YAML::Tiny
module, an incomplete but correct and usable subset of the
functionality, in as little code as possible.