Written by: Peter Hessler <phessler@openbsd.org>
Marathon EVIL is a scenario for Marathon Infinity, originally released
in 1997. It features 17 new solo levels; 22 new network maps; and new
enemies, sounds, textures and weapons. A chief difficulty is the low
amounts of ammunition available, resulting in a greater reliance on
melee weapons, such as fists or staff.

OK merdely@