Written by: Bernd Ahlers <bernd@openbsd.org>
Initial import of ruby-rack-test-0.3.0.

Rack::Test is a small, simple testing API for Rack apps.
It can be used on its own or as a reusable starting point
for Web frameworks and testing libraries to build on.
Most of its initial functionality is an extraction of
Merb 1.0's request helpers feature.