Written by: Ariane van der Steldt <ariane@openbsd.org>
Create libwine.so.0.0 instead of libwine.so.1.0.
Don't symlink libwine.so.1 or libwine.so.
Set WINEPREFIX for regress.
Configure flags: s/without-also/without-alsa/, remove --without-esd, since
that is handled by esd flavor.

(Port is still in wip status, this just makes the port package properly and
exposes the general failure of the regress in this port.)

'Its good if wine.0.0. is created. Let's ask porters.' pirofti@