Written by: Landry Breuil <landry@openbsd.org>
Import Quantum GIS 1.0.2:
QGIS aims to be an easy to use GIS, providing common functions and
features. The initial goal was to provide a GIS data viewer, but now
it also permits to edit, create and analyze lots of spatial datas.
QGIS supports a number of raster and vector data formats, with new
support easily added using the plugin architecture.

with some tweaks by and ok ajacoutot@