Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
Import avahi-0.6.25

Avahi is a free, standards-compliant implementation of the Zeroconf
protocols Multicast DNS and DNS-SD.
It is a system which facilitates service discovery on a local network.
This means that you can plug your laptop or computer into a network and
instantly be able to view other people who you can chat with, find
printers to print to or find files being shared.  This kind of
technology is already found in Apple MacOS X (branded Rendezvous,
Bonjour and sometimes Zeroconf) and is very convenient.

Completely reworked from an old submission from jasper@
Based on FreeBSD and pkgsrc.

Some successful test from marco@, thanks ;-)

It will probably need a bit more love in the next few days.
ok landry@ jasper@