Written by: Landry Breuil <landry@openbsd.org>
Import mapserver 5.4.0:
MapServer is a system for developing web-based GIS applications.
The basic system consists of a CGI program that can be configured to
respond to a variety of spatial requests like making maps, scalebars,
and point, area and feature queries.
This port is splitted in 4 packages: -main contains the mapserv CGI,
-php contains the PHP Mapscript extension, -perl contains Mapscript perl
bindings and -utils contains various utilities and examples of use.
ok ajacoutot@