Written by: Landry Breuil <landry@openbsd.org>
Import keynav:
keynav is a piece of an on-going experiment to make pointer-driven
interfaces easier and faster for users to operate. It lets you move the
pointer quickly to most points on the screen with only a few key
strokes.

Ported because i was too bored writing and reading perl.