Written by: Robert Nagy <robert@openbsd.org>
import ConsoleKit-0.3.0:
ConsoleKit is a framework for defining and tracking users, login
sessions, and seats.  The primary motivations for this framework are to
facilitate fast-user-switching and multi-seat capabilities, and to
enable more sophisticated policy decisions for desktop sessions.

this is not hooked up to the build just yet, just like policykit