Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
Import z-push-1.2.1

Z-push is an implementation of the ActiveSync protocol, which is used
'over-the-air' for multi platform ActiveSync devices, including Windows
Mobile, Ericsson and Nokia phones. With Z-push any groupware can be
connected and synced with these devices.

based on an original port by Gavin Norman.
successfully tested by and ok fkr@