Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
Import hunspell-1.2.8

Hunspell is a spell checker and morphological analyzer library and
program designed for languages with rich morphology and complex word
compounding or character encoding.

Note that this is not to be considered as an aspell replacement just
yet. We install no hunspell dictionnaries for now but use the ones from
mozilla.

Reworked from an original port by Edd Barrett (maintainer).
Tested by sthen@ in a bulk, thanks!
ok sthen@