Written by: Stuart Henderson <sthen@openbsd.org>
the main powerdns binary gets linked with mysqlclient; actually only
the mysql module should be linking with it. work-around this problem
for now by adding a LIB_DEPENDS-main on mysqlclient. this ensures that
people pkg_add'ing have the relevant library available. no functional
change for existing users.

nobody screamed when I sent it to ports@, and it's better than the
current state, but still this is an incorrect fix to unbreak the
package, it should be fixed properly.