Written by: Paul de Weerd <weerd@openbsd.org>
Update to 0.8. This update adds support for receive mode and other
nice features. Of note for OpenBSD users especially is the -z option,
which allows setting the interface description to what was received
from the connected switch.

Tested on amd64, sparc64 and vax.

With suggestions from Brad and sthen@. OK sthen.