Written by: Michael Erdely <merdely@openbsd.org>
Initial import of IP::Country

This includes IP::Country::Fast which SpamAssassin can use to add a
header indicating which country the relay servers are possibly in.

From DESCR:
Finding the home country of a client using only the IP address can be
difficult. Looking up the domain name associated with that address can
provide some help, but many IP address are not reverse mapped to any
useful domain, and the most common domain (.com) offers no help when
looking for country.

This module comes bundled with a database of countries where various IP
addresses have been assigned. Although the country of assignment will
probably be the country associated with a large ISP rather than the
client herself, this is probably good enough for most log analysis
applications, and under test has proved to be as accurate as reverse-DNS
and WHOIS lookup.

ok giovanni@,sthen@