Written by: Landry Breuil <landry@openbsd.org>
Import yokadi 0.10.0.

Yokadi is a command-line oriented, SQLite powered, TODO list tool. It
helps you organize all the things you have to do and you must not
forget. In Yokadi you manage projects, which contains tasks. At the
minimum, a task has a title, but it can also have a description, a due
date, an urgency or keywords.

ok pea@