Written by: Landry Breuil <landry@openbsd.org>
Import thunar-archive 0.2.4.
The thunar-archive-plugin is a plugin for the Thunar File Manager, which
adds archive operations to the file context menus. Using this plugin you
will be able to extract and create archive files from within Thunar using
an external archive handler. So far, xarchiver, file-roller and ark are
supported.

looks good to ajacoutot@