Written by: Stuart Henderson <sthen@openbsd.org>
import the portable version of netbsd's iSCSI target. port req'd by dlg@

iSCSI is an IETF standard (RFC 3720) for remote access to block-level
storage.  It can be thought of as similar to NFS, except that an NFS
server exports files; the iSCSI target exports blocks to the iSCSI
initiators, which are the clients.

To set up the target, you need to edit the /etc/iscsi/targets file.
It has a certain layout, to provide a means of (a) mirroring and (b)
combining multiple areas to present one large contiguous area of
storage.  This can be multiply-layered.