Written by: Bernd Ahlers <bernd@openbsd.org>
Initial import of ruby-passenger-2.2.4.

Phusion Passenger is an Nginx module, which makes deploying Ruby
and Ruby on Rails applications on Nginx a breeze. It follows the
usual Ruby on Rails conventions, such as "Don't-Repeat-Yourself"
and ease of setup, while at the same time providing enough
flexibility.

Tested by many.

With hints from Jeremy Evans and msf@.