Written by: Jasper Lievisse Adriaanse <jasper@openbsd.org>
import py-kiwi 1.9.26

Kiwi consists of a set of classes and wrappers for PyGTK that were
developed to provide a sort of framework for applications. Fully
object-oriented, and roughly Smalltalk's MVC, Kiwi provides a simple,
practical way to build forms, windows and widgets that transparently
access and display your object data.  Kiwi was primarily designed to
make implementing the UI for Stoq easier, and it is released under the
LGPL. Kiwi is inspired by Allen Holub's Visual Proxy.