Written by: Stuart Henderson <sthen@openbsd.org>
import p5-Mail-GnuPG, tweaked from a submission by Toni Mueller.

This module implements the creation and processing of GnuPG signed
and encrypted emails. It provides functions to create PGP/MIME-type
messages, and to check signatures. You need to have a different
package to actually send or receive the message, though.