Written by: Pierre-Emmanuel Andre <pea@openbsd.org>
Import gtkdatabox

A widget to display large sets of data as dots (or lines). Different
data sets will have different colors, you may change colors and style of
the data The widget gives the possibility to zoom into and out of the
data and to scroll around. It also offers a grid like on an
oscilloscope.

With tweaks from landry@
ok landry@