Written by: Landry Breuil <landry@openbsd.org>
Import pigment 0.3.17:
Pigment is a library designed to easily build user interfaces for
embedded multimedia. Its design allows to use it on several platforms,
thanks to a plugin system allowing to choose the underlying graphical
API. Note that it heavily uses OpenGL, so you need a decent hardware
support for it.
Thanks to those who support hackathons, i wouldn't have been able to
spend >1 day on it otherwise.
ok ajacoutot@