Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
Major update to gnome-session-2.28.0... with some part of 2.26.2 because we
do not support the *-kit thingies.
There is a small regression wrt to shutdown/reboot which I'll work on in
the next days.
Also, this port is not a meta-port anymore, if you want a basic functionnal
GNOME desktop, make sure you read the README.OpenBSD file (as agreed with
my other half jasper@).

so many cute animals got injured while working on this update, I lost count...
(sorry kiki...)