Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
Display the usual icons under the System menu by default.
This change was introduced for GNOME 2.28.0 for whatever reason
(performance?), but for now not all applications follow this and it makes
several parts of the desktop/menus look weird to say the least.

So we use previous behaviour because we like beauty ;-)