Written by: Bernd Ahlers <bernd@openbsd.org>
Initial import of graph-1.43.

Grap is a language for typesetting graphs specified and first
implemented by Brian Kernighan and Jon Bentley at Bell Labs. It is an
expressive language for describing graphs and incorporating them in
typeset documents. It is implemented as a preprocessor to Kernigan's pic
language for describing languages, so any system that can use pic can
use grap. For sure, TeX and groff can use it.

From maintainer Hermann Gottschalk.