Written by: Stuart Henderson <sthen@openbsd.org>
import www/drupal6/theme-acquia-marina

Acquia Marina is an advanced fixed-width theme, supporting 1, 2 or
3-column layouts with 15 collapsible block regions. It has an optional
drop-down primary links menu and advanced theme settings to customize
without coding. Icons for core and Views blocks are included.

It is cross-browser tested (in IE6/7, Opera, Safari, and Firefox)
and has a high attention to detail.