Written by: Landry Breuil <landry@openbsd.org>
Add a pair of patches to resindvd plugin to set DEFAULT_DEVICE to
/dev/rcd0c instead of /dev/dvd.. no better workaround atm, but at
least gstreamer-enabled apps have more chances to play dvds.
(btw, you have to be in operator group to access the device)
Note that gstreamer-ugly also provides some more or less working
plugins for dvd playing, but resindvd is supposed to be the way to go.

ok ajacoutot@ with some grumblings