Written by: Marc Espie <espie@openbsd.org>
less confusing version number (okay, we still have always-update, but let's
try harder not to go back).
put more packages we know to be gone in the base system list.
display a more helpful message in case we don't find base packages (say
we're not removing them since we can't find the file in the base system)