Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
Use setrlimit to set a correct value for maxfilesperproc.

Disable the kqueue(2) backend for now (by setting max_open_files to the
minimum) as there are some weird corner cases where it doesn't work and
hangs gam_server. I'm still on that but for the time beeing we want this
to work in every situations.