Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
Remove the -gimp subpackage. It had a meaning when gimp did not provide
a sensible printing interface but this is not the case anymore. This
also allows to remove gutenprintui2 and gtk+2 dependency.

Rebuild -foomatic against the latest foomatic-db-engine.
The -foomatic package run_depends on foomatic-db. Move some files from
-main to -foomatic.