Written by: Stuart Henderson <sthen@openbsd.org>
Update to 3.6.16.1; a simple but important fix (and needed to update
Firefox). Mainline sqlite has moved beyond this but involves some changes
to the build infrastructure which need to be looked at separately.

Since the distfile is no longer available upstream, this is provided
as a patch to the 3.6.16 distfile and is taken from upstream's SCM.

Requested by naddy@.